LayoutTests/imported/w3c/web-platform-tests/svg/path/property/test_style_flush_on_dom_api_with_d_property-expected.txt


FAIL getTotalLength() with d property assert_equals: the initial d expected (string) "path(\"M 0 0\")" but got (undefined) undefined
FAIL getPointAtLength() with d property assert_equals: the initial d expected (string) "path(\"M 0 0\")" but got (undefined) undefined
FAIL isPointInFill() with d property assert_equals: the initial d expected (string) "path(\"M 0 0\")" but got (undefined) undefined
FAIL isPointInStroke() with d property assert_equals: the initial d expected (string) "path(\"M 0 0\")" but got (undefined) undefined
